Pull up as much as possible into common
