+++ eithernet/src/commonMain/kotlin/com/slack/eithernet/ApiException.kt
