+++ eithernet/src/commonMain/kotlin/com/slack/eithernet/ResultType.kt
@@ -20,7 +20,7 @@ import kotlin.reflect.KClass
 
 /**
  * Represents a [java.lang.reflect.Type] via its components. Retrieve it from Retrofit annotations
- * via [errorType] and piece this back into a real instance via [toType].
+ * via [errorType] and piece this back into a real instance via `ResultType.toType()`.
  *
  * This API should be considered read-only.
  */
