+++ eithernet/src/commonMain/kotlin/com/slack/eithernet/StatusCode.kt
