+++ eithernet/src/commonMain/kotlin/com/slack/eithernet/retries.kt
@@ -23,29 +23,6 @@ import kotlin.time.Duration.Companion.seconds
 import kotlin.time.times
 import kotlinx.coroutines.delay
 
-@Suppress("LongParameterList")
-@Deprecated("Left for ABI compatibility", level = DeprecationLevel.HIDDEN)
-public suspend fun <T : Any, E : Any> retryWithExponentialBackoff(
-  maxAttempts: Int = 3,
-  initialDelay: Duration = 500.milliseconds,
-  delayFactor: Double = 2.0,
-  maxDelay: Duration = 10.seconds,
-  jitterFactor: Double = 0.25,
-  onFailure: ((failure: ApiResult.Failure<E>) -> Unit)? = null,
-  block: suspend () -> ApiResult<T, E>,
-): ApiResult<T, E> {
-  return retryWithExponentialBackoff(
-    maxAttempts = maxAttempts,
-    initialDelay = delayFactor * initialDelay,
-    delayFactor = delayFactor,
-    maxDelay = maxDelay,
-    jitterFactor = jitterFactor,
-    onFailure = onFailure,
-    shouldRetry = { true },
-    block = block,
-  )
-}
-
 /**
  * Retries a [block] of code with exponential backoff.
  *
